Add A/B implementation.
This CL adds the avb_ab_flow() routine which provides A/B seletion
including managing rollback indexes.

A/B metadata is stored in the 'misc' partition in the |slot_suffix|
field using a format private to libavb - see bootable/recovery/bootloader.h
for more details.

A/B metadata integrity is provided by a simple magic marker and a CRC-32
checksum. If invalid A/B metadata is detected, the behavior is to reset
the A/B metadata to a known state where both slots are given seven boot
tries.

Future CLs will provide a boot_control HAL implementation compatible
with the A/B implementation in libavb as well as a avbtool subcommand to
initialize metadata on a 'misc' image.

Also factored out the test-side AvbOps into a FakeAvbOps class and put
it in its own file.

BUG=31264229
TEST=New unit tests + all unit tests pass.

Change-Id: I594ea4173a051ecb72636058440372ff1ca5855b