Minor fixes to make the boot_control HAL work on real hardware.
Drop -fvisibility=hidden since we actually need to access avb_* symbols
from the libavb library

Add support for the isSlotMarkedSuccessful() function in the
boot_control HAL implementation - it wasn't in boot_control_copy which
the implementation was initially based on.

Create a symlink such that the boot_control HAL is actually loaded by
default regardless of what hardware is used.

Tested the boot control HAL on a real DUT (well, qemu with a full disk
image which is very close) using the bootctl(1) program - everything now
works as expected.

Bug: 31264229
Change-Id: If1edc7823f8d6341e2fbe7e9030d261ec359c87d
Test: Manually tested on brillo_uefi_x86_64 board. Unit tests pass.