Include full vbmeta digest on the kernel cmdline.
While manually sanity-checking the androidboot.vbmeta.digest value (by
running sha256sum on the DUT) I discovered that only half the digest was
appended to the kernel command-line. While we already had unit tests for
checking this, the data hardcoded in the unit tests was also half the
size.

Here's the output from the DUT for manually testing the fix:

 # getprop |grep vbmeta
 [ro.boot.vbmeta.device_state]: [unlocked]
 [ro.boot.vbmeta.digest]: [f75dc1643b48d19696d001b0f6ef90440ef2df43253c00c4675f53fa70b3ab64]
 [ro.boot.vbmeta.hash_alg]: [sha256]
 [ro.boot.vbmeta.size]: [2944]

 # dd if=/dev/block/pci/pci0000\:00/0000\:00\:01.1/by-name/vbmeta_a bs=1 count=2944 |sha256sum
 f75dc1643b48d19696d001b0f6ef90440ef2df43253c00c4675f53fa70b3ab64  -

Bug: None
Change-Id: I60ad123b9abaa5f271ffd256964eb63acfb3807d
Test: Manually tested on DUT, see above
Test: Updated unit tests and all unit tests pass.