Update patch set 1
Patch Set 1: Verified+1 Autosubmit+1

Hi - sorry for another large-ish CL and thanks in advance for taking a look!

The good news is that I'm making good progress on all these interesting and easy-to-forget corner-cases. Next up (later this week) is a CL for introducing a field in VBMeta so 'adb disable-verity' works as expected e.g. to set up the rootfs without verify. This CL is doing a lot of the footwork to enable that.

Patch-set: 1
Label: Autosubmit=+1
Label: Verified=+1