Update patch set 2
Patch Set 2:

> Why is this needed?  Can't adb disable-verity just recreate vbmeta with a hardcoded root=PARTUUID=blah?

That's another (more direct) way to implement it, sure. But I don't like it, for several reasons.

First I don't think it's very elegant since it's essentially encoding knowledge into 'adb disable-verity' about how kernels are booted and what the vbmeta data structure looks like. You also can't just recreating vbmeta - you need to include the other descriptors as well. It would also mean reimplementing a lot of avbtool in adbd which is tricky given that avbtool is in python (and runs at build time) and adbd is C++.

Second, it's a lot of easier to write tests with it being this way .. it's literally just writing a byte to the vbmeta partition in a given place ... so easy you can even use busybox and the sh on the device to do it cf. the commit message. Minimizing dependencies for testing is a good thing as I'm sure you would agree with.

Patch-set: 2