Update patch set 2
Patch Set 2:

> Doesn't that make sense rather than encumber the _bootloader_ with this support and making the vbmeta larger by keeping two commandlines?

The extra storage and code is a non-issue.

> If we always generated an unsigned disable-verity vbmeta and put it in the image, then you could just use 'dd' to switch over.

From a build-system point of view it would add complications (modifying the image we're also adding dm-verity to). Not a path I want to go down.

> I'm still not sold that this belongs in a reference bootloader that may be tweaked by integrators in unexpected ways.
> Am I being crazy?

I really don't see the problem, sorry. The two command-lines are added automatically by avbtool and processed automatically in the bowels of libavb. Both are far away from any integration points.

In fact the way it's done is simpler than what you propose because all the logic is centralized in a single place - avbtool. Also, having it as a flag in the struct allows you to easily inspect whether 'adb disable-verity' was ever run. Which is helpful for forensics / debugging / etc.

I appreciate the feedback but aren't we arguing over an implementation detail or am I way off? Happy to talk more on VC!

Patch-set: 2