Update patch set 2
Patch Set 2:

> > Doesn't that make sense rather than encumber the _bootloader_
 > with this support and making the vbmeta larger by keeping two
 > commandlines?
 > 
 > The extra storage and code is a non-issue.

This adds a single int which can be overwritten to bypass verification, right?  That seems not ideal in sensitive code when the branch/value is only needed for development and debugging.  Verified boot code should, imo, be as minimal and single-path focused as possible to avoid exploding combinations of failure states -- accidental or malicious.  I think testing thoroughly catches the accidental, but I don't see the value of that bit of code living in there.

The counter-argument, of course, would be that a well timed attack for replacing the 'unlocked' bit in memory would be just as effective if it occurred before the vbmeta signature verification check. 

On the whole, it just seems easier to have one insecure path through the bootloader than more than one.

 > 
 > > If we always generated an unsigned disable-verity vbmeta and put
 > it in the image, then you could just use 'dd' to switch over.
 > 
 > From a build-system point of view it would add complications
 > (modifying the image we're also adding dm-verity to). Not a path I
 > want to go down.
 
Better than in the bootloader, imo.  It is almost always better to do all the work during build than at runtime, imho.

 > > I'm still not sold that this belongs in a reference bootloader
 > that may be tweaked by integrators in unexpected ways.
 > > Am I being crazy?
 > 
 > I really don't see the problem, sorry. The two command-lines are
 > added automatically by avbtool and processed automatically in the
 > bowels of libavb. Both are far away from any integration points.
 
Both are in the vbmeta though, correct?

 > In fact the way it's done is simpler than what you propose because
 > all the logic is centralized in a single place - avbtool. Also,
 > having it as a flag in the struct allows you to easily inspect
 > whether 'adb disable-verity' was ever run. Which is helpful for
 > forensics / debugging / etc.

If vbmeta is signed by a devkey or an unknown key, it is easy to inspect as well -- beyond the kernel cmdline.

 > 
 > I appreciate the feedback but aren't we arguing over an
 > implementation detail or am I way off? Happy to talk more on VC!

We are definitely arguing over an implementation detail! It's a code review :) I'm arguing that this code shouldn't exist and you're arguing it is fine to exist.

I'm not sure the convenience is worth it though. Is there something really compelling I am missing?

Patch-set: 2
Label: Verified=0