Update patch set 2
Patch Set 2:

> <flipping bits in memory>

Mmm. I get what you're saying. But. On a practical level, I don't think this changes the situation one iota. I mean, yes, if you have this capability (flipping bits in memory) there are now N+1 ways to affect the bootloader, not just N. But what you are not saying is that N is already huge. I mean, it's almost like saying "don't add more lines of code, they may contain bugs!" ... which, while definitely true, is generally not very helpful advice if you're being tasked with adding a feature.

> It is almost always better to do all the work during build than at runtime, imho.

The *actual* change to the bootloader is minimal, it's basically just the test against VBMeta.flags and CmdlineDescriptor.flags and then use the result of this whether to apply the cmdline or not. That's the amount of code added. (Oh, and also byte-swapping the flags field and rejecting chained partitions with non-zero flags - but that's also a trivial amount of code.)

> I'm not sure the convenience is worth it though. Is there something really compelling I am missing?

You are right that it's a cost/benefit thing. The cost is adding almost no data and almost no code (see above) and the benefits are the following:

- we avoid leaking details about how the system is booted or how the kernel is configured ... useful if we e.g. want to insert an additional mapped device in the stack [1] - we'd just change avbtool to emit the right command-lines for both cases. Yes, you said you get get around this by introducing vbmeta-for-dm-verity-disabled.img in the image but...

- having to insert vbmeta-for-dm-verity-disabled.img in the image (at say /system/avb) is error-prone and - let's be honest - just an eyesore. And build/core/Makefile + friends (python scripts) are already pretty complicated and this would make it worse. Worse, what if vbmeta-the-partition and /system/avb/vbmeta-for-dm-verity-disabled.img gets out of sync for whatever reason? It's just going to complicate debugging and a non-trivial amount of people can't get work done because 'adb disable-verity' is now broken. Much easier to just have 'adb disable-verity' write a single byte to the vbmeta partition and we know this works because we've already unit- *and* integration-tested libavb (including this feature).

- having a HASHTREE_DISABLED flag convey that we *explicitly* support this feature and there's a way to know if the device is in this state. This is useful integrators insofar that it reminds them that this is a legitimate case that they need to test. Additionally there may be a need for integrators to have this explicit information if they e.g. have hardware features to disable writing to the partition (area of the disk) in question before jumping to Android (if HASHTREE_DISABLED is set they know not to do this). [2]

I guess my main concern is that what you are proposing is... I want to say "layering violation" (knowing very well that it's a dirty word to use)... yes, this CL adds an additional branching point and a slightly larger vbmeta partition... which, yes, is a cost at least in some extreme sense of the word. In my view it's a cost well-worth it because of the reasons above.

    David

[1] : in this case we'd have

 dm-some_other_mapping_target
 dm-verity

for the dm-verity-is-enabled case and

 dm-some_other_mapping_target

for the dm-verity-is-disabled case.

[2] : there could be a world where updates are verified and applied not by update_engine in Android... but by something at the bootloader/preboot level... yes, crazy-talk, I know, but it could be this isn't so crazy after all in some version of the future.

Patch-set: 2