Update patch set 2
Patch Set 2:

(Btw I just remembered: the other thing is that the bootloader is not the sole user of the HASHTREE_DISABLED flag - we need it in fs_mgr as well to skip setting up dm-verity for non-rootfs partitions. Without this flag we'd need to use some kind of hacky heuristic - parsing /proc/cmdline or checking if dm-verity is used on rootfs - to infer what to do. It's much cleaner having the flag there explicitly.)

Patch-set: 2