Update patch set 3
Patch Set 3:

> Will fs_mgr re-verify the vbmeta signature or will it blindly trust it since it booted?

I don't think having fs_mgr blindly trust the data is a good idea (defense in depth, etc.) - the best would be if it can verify the data is authentic.

Anyway, see https://android-review.googlesource.com/#/c/307283/ for some discussion. The tl;dr is that androidboot.vbmeta.digest is set - by the bootloader - to the digest of all the vbmeta structs combined (before the CL it is set only for the top-level vbmeta struct).

This way fs_mgr - or any user-space app - can simply call avb_slot_verify() and then at the end calculate its own digest (by hashing all elements in AvbSlotVerifyData.vbmeta_images[] array) and compare it to androidboot.vbmeta.digest. If it matches, _all_ vbmeta structs - not just the one from the vbmeta partition - in AvbSlotVerifyData can be trusted (as much as /proc/cmdline can be trusted). Having established authenticity, fs_mgr can now simple examine descriptors in all vbmeta structs and do its thing.

The nice thing about this setup is that user space does not need read-access to neither the key used to sign top-level vbmeta nor any of the roll-back indexes. I think this is pretty important (restrict access to the SE as much as possible).

The other nice thing (about the CL) is that fs_mgr - or any user-space app - can re-use avb_slot_verify() and not have to reimplement walking the AVB trees itself.

> Additionally, don't we populate some props from the kernel commandline to communicate to fsmgr about the boot state?
> Why would we make another mechanism? Wouldn't it need to update each other partitions vbmeta to have the hash tree disabled rather than just having a global prop that says "disable verity" or whatever?

But how would the bootloader know that dm-verity is to be disabled? It would have to sniff the command-line. This is essentially just moving the heuristic to another place (and I'd like to avoid as many string operations in libavb as I can for security reasons).

> fs_mgr seems orthogonal but I'm not thinking the most clearly today ;)

fs_mgr is just one example - all this applies to any user-space app...

Patch-set: 3