Update patch set 3
Patch Set 3:

> Seems like a bad trick to teach the bootloader? :)

Not at all! As mentioned in my Nov 28 3:14pm comment (no comment numbers in Gerrit .. grr) above the boot loader may actually need to know if it should set set the LUN with non-/data partitions to read-only (or part of the LUN) before jumping to Android (for hypothetical cases where the system updater runs outside Android). And the easiest way for the boot loader to do this without breaking 'adb disable-verity' is to check a HASHTREE_DISABLED flag in the vbmeta struct.

Yes, it's possible to implement "Add way to disable dm-verity allowing rootfs to be writable." in more than one way, each with their pros and cons. I still maintain that the solution in this CL is the cleanest one all things considered (attack surface, additional data required, future-proof, robustness, build complexity, etc).

From a 10,000 feet point of view I think this comes down to accepting the fact that the bootloader *need* to have *some* developer features built in. Case in point is https://android-review.googlesource.com/#/c/304672/ that adds a feature to support booting unverified (but valid) images when in UNLOCKED mode and this adds quite some complexity for example the introduction of the |allow_verification_error| parameter and AVB_AB_FLOW_RESULT_OK_WITH_VERIFICATION_ERROR return value. Without this it's not possible for a bootloader to distinguish between invalid, valid but unverified, and verified images. That CL adds more complexity than I want but it does enable a developer feature that we need. It's a tradeoff!

I feel like this particular CL is right up that alley. Yes, we trade off a tiny tiny bit of complexity (much simpler than the "allow ops on unlocked devices" CL just mentioned) and 64 extra bytes (the string "root=PARTUUID=$(ANDROID_SYSTEM_PARTUUID)" and its KernelCmdLine header struct) in vbmeta.img. In return, we have a very robust way to implement 'adb disable-verity' (just set a single bit in 'vbmeta') which is kinda important since it costs Google money every time this feature does not work for one of its employees. Tradeoffs!

Your suggestion of computing two vbmeta.img files at build-time and storing one of them in /system/avb/vbmeta-for-when-hashtree-is-disabled.img also gets the job done and avoids the tiny-tiny-complexity-above and the 64 bytes of extra storage. In return it complicates the build process and makes 'adb disable-verity' more error-prone. It's simply not worth it. (And it also won't work for the corner-cases where the KernelCmdline descriptor for 'system' is in a chained partition).

(Sorry for sounding like a broken record. I will keep future responses to one paragraph or less!)

Patch-set: 3