avbtool: add support for signing helper.
Right now avbtool simply uses 'openssl' program for signing hashes. A
key is passed via the --key options. It is fine for development and
testing but it is a problem if we are trying to integrate with an existing
singing server.

Intoduce the --signing_helper option which can be used to specify any
external program for signing hashes. The data to sign is fed via STDIN
and the sighed data is returned via STDOUT. If --signing_helper is
specified in a command line, the --key option need only contain a public key.

Bug: 33586278
Test: New unit tests and all unit tests pass.
Change-Id: I74d4b5f076f8a25a29924394fff4708bc1ebf970
Signed-off-by: Dzmitry Yatsushkevich <dmitryya@google.com>