avbtool: add --set_hashtree_disabled_flag option to make_vbmeta_image command.
This enables people to set

 BOARD_AVB_MAKE_VBMETA_IMAGE_ARGS="--set_hashtree_disabled_flag"

in their environment and then the images they build are baked with the
HASHTREE_DISABLED bit set and as a result they don't need to manually
disable dm-verity via e.g. 'adb disable-verity'. Note that this is
already possible through the existing --flags option, e.g.

 BOARD_AVB_MAKE_VBMETA_IMAGE_ARGS="--flags 1"

but it's much nicer to be explicit about it.

Bug: 34124301
Test: Unit tests pass.
Test: Manually tested on UEFI based bootloader.
Change-Id: I75537c614a660362ae68342ce50aa2c09022dc07