Update patch set 1
Patch Set 1:

(I also expect that we can iterate on avbctl as needed... e.g. in the future we could an --slot option to the enable-verity and disable-verity verbs that could assume the values _a/_b/all. And we could add code to verify the signature, the sky is the limit etc. etc.. .... I want to get the basics working first, though.)

Patch-set: 1
Reviewer: David Zeuthen <1076971@85c56323-6fa9-3386-8a01-6480fb634889>