Move enable/disable-verity code from avbctl to libavb_user.
Also build a libavb_user library and make avbctl link with it. Also
add a function to get whether verity is currrently enabled and expose
this through a new get-verity sub-command in avbctl. Make avbctl a bit
smarter and tell the user if verity is already enabled/disabled when
trying to enable/disable it. Also remind the user to reboot if we're
changing the state.

With this change it's possible to make adbd link with libavb_user for
implementing 'adb {enable,disable}-verity' command.

Also stop linking avbctl with the boot_control HAL and just access the
ro.boot.slot_suffix (aka "androidboot.slot_suffix" on the kernel
cmdline). This way external/avb should be usable on N without manually
tweaking it.

Complete AvbOps in libavb_user so it can be used with avb_slot_verify()
in user space. Also add more documentation explaining what this particular
AvbOps implementation is intended to do.

Bug: 34124301
Test: All unit tests pass.
Test: Manually tested on UEFI-based bootloader.
Change-Id: I0b5080ff102c29f2c4c0e92a9737590d450967a9