libavb: Load entire partition if |allow_verification_error| is true.
This is needed to make the common workflow of

 $ fastboot flash boot /path/to/boot.img

work. To do this we need to introduce a new AvbOps operation to get
the partition size. Note that libavb integrators had already had to do
this to implement read_from_partition() when a negative offset is
passed.. so this shouldn't be a lot of extra work.

Also note that since libavb has no stable API this is not a breaking
change so there is no need to bump any version numbers (version
numbers are mostly for on-disk formats) ... in other words, libavb
integrators are expected to re-integrate and re-test their bootloader
code every time they uprev to a newer libavb. In this case they need
to implement a new AvbOps operation.

Since we're introducing a new operation add some pre-condition checks
to avb_slot_verify() to check that all required operations at least
are set.

Add a new unit test for this and also implement it in the
examples/uefi boot loader.

Bug: 37709309
Test: New unit test + all unit tests pass.
Test: Manually tested on UEFI-based bootloader.
Change-Id: Id225af91add2e52167994e80b5b3a788c6909c15