Update AVB docs to include overview, recommendations and best practices.
This information is currently in a separate document which is
confusing when things get out of sync, leading to bitrot and
worse. Instead, just include this information in README.md - this way
it's easy to point to and things don't bitrot!

Emphasize that recommendations and best practices are just that and
the word 'must' should be taken with a grain of salt. It is done this
way such that requirement docs for specific devices can point to a
specific version of README.md and say "the device MUST implementation
all recommendations (except X, Y, and Z) in external/avb/README.md".

Test: Renders fine in Markdown editor.
Test: All unit tests pass.
Bug: None
Change-Id: Iaae80e021180c1a18db8cdbd479e04c98f4c7d33