Update patch set 1
Patch Set 1:

In addition to this we'll also need a fs_mgr CL (easy, already got one) and the dm-drivers/md/verity-avb.c pieces (harder, got it working but right now cleaning it up).

Also, don't be intimidated by this CL - all it does is to pass the error-mode from the bootloader to the HLOS.. e.g. it maps a AvbHashtreeErrorMode code to androidboot.veritymode + androidboot.invalidate_on_error.

Tests and documentation is why this CL is so big.

Patch-set: 1
Reviewer: David Zeuthen <1076971@85c56323-6fa9-3386-8a01-6480fb634889>