Update patch set 1
Patch Set 1:

Not quite ready, but pushing in case anyone else is also thinking about working on this.

I want to convert init to .bp and that requires converting libfs_mgr, which requires converting libavb and so I did (most) of this file in the process.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>