avbtool: Make verify_image sub-command verify hash- and hastree descriptors.
Up until now 'avbtool verify_image' only checks that the vbmeta struct
is signed with the embedded public key. While this is useful it would
be more useful if we also checked that the digests in the hash- and
hashtree descriptors are correct.

The complication with this is that we need to load other image files
and we don't really know what file to look for - all we have is a
partition name. The way we solve is is that we look in the same
directory as the given image and also use the file extension of the
given image. E.g. if you use --image /path/to/vbmeta.img and we
process a hash descriptor for 'boot' it will make us look for
/path/to/boot.img.

Here's an example of a garden-vanilla setup where all integrity-data
is stored in vbmeta.img:

 $ ./avbtool verify_image --image ${ANDROID_PRODUCT_OUT}/vbmeta.img
 Verifying image /ssd/android/aosp/out/target/product/uefi_x86_64/vbmeta.img using embedded public key
 Successfully verified SHA256_RSA4096 vbmeta struct in /ssd/android/aosp/out/target/product/uefi_x86_64/vbmeta.img
 Successfully verified sha256 hash of /ssd/android/aosp/out/target/product/uefi_x86_64/boot.img for image of 10543104 bytes
 Successfully verified sha1 hashtree of /ssd/android/aosp/out/target/product/uefi_x86_64/system.img for image of 1065213952 bytes
 Successfully verified sha1 hashtree of /ssd/android/aosp/out/target/product/uefi_x86_64/vendor.img for image of 264114176 bytes
 $ echo $?
 0

Here's an example where we want to ensure vbmeta.img is signed with a
given key. It fails because it's signed by the default test key which
is test/data/testkey_rsa4096.pem

 $ ./avbtool verify_image --image ${ANDROID_PRODUCT_OUT}/vbmeta.img --key test/data/testkey_rsa2048.pem
 Verifying image /ssd/android/aosp/out/target/product/uefi_x86_64/vbmeta.img using key at test/data/testkey_rsa2048.pem
 ./avbtool: Embedded public key does not match given key.
 $ echo $?
 1

If we pass that key verification succeeds

 $ ./avbtool verify_image --image ${ANDROID_PRODUCT_OUT}/vbmeta.img --key test/data/testkey_rsa4096.pem
 Verifying image /ssd/android/aosp/out/target/product/uefi_x86_64/vbmeta.img using key at test/data/testkey_rsa4096.pem
 Successfully verified SHA256_RSA4096 vbmeta struct in /ssd/android/aosp/out/target/product/uefi_x86_64/vbmeta.img
 Successfully verified sha256 hash of /ssd/android/aosp/out/target/product/uefi_x86_64/boot.img for image of 10543104 bytes
 Successfully verified sha1 hashtree of /ssd/android/aosp/out/target/product/uefi_x86_64/system.img for image of 1065213952 bytes
 Successfully verified sha1 hashtree of /ssd/android/aosp/out/target/product/uefi_x86_64/vendor.img for image of 264114176 bytes
 $ echo $?
 0

Also, verification still work for chained partitions and provided the
filename matches the partition name (modulo extension) then
hash/hashtree verification machinery also kicks in:

 $ ./avbtool verify_image --image ${ANDROID_PRODUCT_OUT}/system.img
 Verifying image /ssd/android/aosp/out/target/product/uefi_x86_64/system.img using embedded public key
 Successfully verified footer and SHA256_RSA4096 vbmeta struct in /ssd/android/aosp/out/target/product/uefi_x86_64/system.img
 Successfully verified sha1 hashtree of /ssd/android/aosp/out/target/product/uefi_x86_64/system.img for image of 1065213952 bytes
 $ echo $?
 0

Also make 'avbtool verify_image' a lot more verbose so it's easier to
see what's going on.

Bug: 62038348
Test: DO NOT MERGE - need to write unit tests
Change-Id: Ic6cd48c9ef9c8d97f6427104466bdc1822fc54aa