avbtool: Add --print_required_libavb_version option.
This can be used with make_vbmeta_image, add_hash_footer, and
add_hashtree_footer to figure out what will be put in the
required_libavb_version_{major, minor} fields in the VBMeta
struct. This is useful if you also want to store this value in a file
inside system.img before adding a hashtree and integrity data to it.

Right now "1.0" will be printed out since we're still at 1.0. In the
future this may get more complicated.

As with --calc_max_image_size the caller is supposed to pass all the
same options (except --output and --image) that he would pass when not
using the option. This is because options will influence what minimum
libavb version will be required - for example, if using an algorithm
introduced post-1.0 this would be expressed with the --algorithm
option.

Bug: 62077941
Test: New unit test and all unit tests pass.
Change-Id: I97a26f5b51eb4a8c63ca7903a406b174eb88bacb