Remove libavb_ab and bootctrl.avb code.
This code was already marked as experimental in anticipation of being
removed in the future. The reason for removing this code is twofold:

 - Its policy was optimized for devices without a display with
   e.g. automatic fallback to the other slot if a slot fails to
   boot. Since most A/B stacks in Android devices don't work this
   way this code is confusing.

 - There are no known active users, no good test coverage for the
   bootctrl.avb code, and no plans to use it.

If someone is already using the libavb_ab and bootctrl.avb code they
can continue to do so with minor changes - just use |ab_user_data|
instead |ab_ops| in the AvbOps data structure.

Also port the UEFI-based boot loader in examples/uefi/ to use libavb
and hard-code slot A for now. Leave a TODO about how it would be nice
to have a super-slim A/B stack there to point out integration points.

Test: All unit tests pass.
Test: UEFI-based boot loader compiles.
Change-Id: Id5e090a2048076d36ccca2e1c4cb55e226b8b43d